refactor(www): move the styling engine to lib/styles
modules/core was the one non-feature in modules/ — it's the styling
engine (DesignSystemProvider, createStyles, emitCss), imported by ~44
registry styles.ts files plus the showcase and preview, with no route of
its own. Move it to lib/styles so modules/ holds only route-features
(create, docs, marketing).

Rewrite all @/modules/core/styles imports to @/lib/styles, repoint the
v0 page template + materialize harness, and regenerate showcase-bundle.ts
(it ships lib/styles.tsx now). styles.tsx carries no Tailwind classes, so
no @source change is needed.
