+++ www/src/registry/ui/accordion/styles.ts
@@ -1,4 +1,4 @@
-import { createStyles } from '@/modules/core/styles'
+import { createStyles } from '@/lib/styles'
 
 import accordionMeta from './meta'
