+++ www/src/registry/ui/dialog/styles.ts
@@ -1,4 +1,4 @@
-import { createStyles } from '@/modules/core/styles'
+import { createStyles } from '@/lib/styles'
 
 import dialogMeta from './meta'
