+++ www/src/registry/ui/drawer/styles.ts
@@ -1,4 +1,4 @@
-import { createStyles } from '@/modules/core/styles'
+import { createStyles } from '@/lib/styles'
 
 import drawerMeta from './meta'
