+++ www/src/registry/ui/skeleton/styles.ts
@@ -1,4 +1,4 @@
-import { createStyles } from '@/modules/core/styles'
+import { createStyles } from '@/lib/styles'
 
 import skeletonMeta from './meta'
