+++ www/src/registry/ui/toggle-button-group/styles.ts
@@ -1,4 +1,4 @@
-import { createStyles } from '@/modules/core/styles'
+import { createStyles } from '@/lib/styles'
 
 import toggleButtonGroupMeta from './meta'
