+++ www/src/registry/ui/tooltip/styles.ts
@@ -1,4 +1,4 @@
-import { createStyles } from '@/modules/core/styles'
+import { createStyles } from '@/lib/styles'
 
 import tooltipMeta from './meta'
