+++ www/src/routes/r/showcase-bundle.tsx
@@ -145,7 +145,7 @@ function buildPageTsx(preset: PublishPreset): string {
 import "./globals.css";
 
 import Cards from "../modules/marketing/cards";
-import { DesignSystemProvider } from "../modules/core/styles";
+import { DesignSystemProvider } from "../lib/styles";
 
 export default function Page() {
 \treturn (
